Attempt to get build & package working
