+++ .github/workflows/build.yml
@@ -1,0 +1,37 @@
+name: Build node_printer
+permissions:
+  contents: write
+  id-token: write
+on:
+  push:
+    branches: ["main"]
+  workflow_dispatch:
+jobs:
+  prebuild:
+    name: Prebuild Modules
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v4
+      - uses: ilammy/msvc-dev-cmd@v1
+        if: matrix.os == 'windows-latest'
+      - uses: actions/setup-node@v4
+        with:
+          node-version: ${{ vars.NODE_BUILD_VERSION }}
+          cache: "npm"
+      - name: Prebuild
+        run: |
+          prebuild_targets="-t $(echo "$NODE_TARGET_VERSIONS" | tr '\n' ' -t ')"
+          npx prebuildify $prebuild_targets --no-napi --strip
+        shell: sh
+        env:
+          NODE_TARGET_VERSIONS: ${{ vars.NODE_TARGET_VERSIONS }}
+      - name: Upload Prebuilds
+        uses: actions/upload-artifact@v4
+        with:
+          name: prebuilds-${{ matrix.os }}
+          path: prebuilds/
+          if-no-files-found: error
+          retention-days: 30
