--- .travis.yml
@@ -1,44 +1,0 @@
-language: node_js
-node_js:
-  - 14
-
-branches:
-  only:
-  - master
-  
-env:
-  global:
-    - JOBS=4
-
-cache:
-  npm: true
-  directories:
-    - $HOME/.npm/_prebuilds
-
-before_script:
-- if [[ $TRAVIS_OS_NAME == "linux" ]]; then export CXX='g++-4.8'; fi
-- git config --global user.name "Ion Lupascu"
-- git config --global user.email ionlupascu@gmail.com
-
-script:
-  - npx prebuild --strip --all -r electron --arch ia32 -u $PREBUILD_TOKEN
-  - npx prebuild --strip --all -r electron --arch x64 -u $PREBUILD_TOKEN
-  - npx prebuild --strip --all --arch ia32 -u $PREBUILD_TOKEN
-  - npx prebuild --strip --all --arch x64 -u $PREBUILD_TOKEN
-
-os:
-  - linux
-  - osx
-
-addons:
-  apt:
-    sources:
-      - ubuntu-toolchain-r-test
-    packages:
-      - g++-4.8-multilib
-      - gcc-4.8-multilib
-      - libcups2-dev
-      - libcups2-dev:i386
-      - libc6-dev-i386
-      - linux-libc-dev 
-      - linux-libc-dev:i386
