--- appveyor.yml
@@ -1,50 +1,0 @@
-# appveyor file
-
-# build version format
-version: "{build}"
-
-platform: 
-  - x64
-  - x86
-
-image: Visual Studio 2017
-
-# Set a known clone folder
-clone_folder: c:\p
-
-# fix lineendings in Windows
-init:
-  - git config --global core.autocrlf input
-  - git config --global user.name "Ion Lupascu"
-  - git config --global user.email ionlupascu@gmail.com
-
-environment: 
-  UTO:
-    secure: <secure_token_generated_from_appveyor>
-
-matrix:
-  fast_finish: true
-
-install:
-  - ps: Install-Product node 14 $env:PLATFORM
-  - cmd: SET PATH=c:\python27;%PATH%
-  - cmd: SET JOBS=4
-  - cmd: SET BUILD_ONLY=true
-  - cmd: npm ci --build-from-source
-  - cmd: if %PLATFORM% EQU x86 (SET TP=ia32) else (set TP=x64)
-  - cmd: npx prebuild --strip -r electron --all --arch %TP% -u %UTO%
-  - cmd: npx prebuild --strip -r node-webkit --all --arch %TP% -u %UTO%
-  - cmd: npx prebuild --strip --all --arch %TP% -u %UTO%
-test_script:
-  - node --version
-  - npm --version
-
-build: off
-
-branches:
-  only:
-    - master
-
-cache:
-  - '%APPDATA%\npm-cache'
-  - '%USERPROFILE%\.node-gyp'
