+++ binding.gyp
@@ -1,26 +1,11 @@
 {
-  "variables": {
-    "module_name%": "node_printer",
-    "module_path%": "lib"
-  },
   'targets': [
     {
-      "target_name": "action_after_build",
-      "type": "none",
-      "dependencies": [ "<(module_name)" ],
-      "copies": [
-        {
-          "files": [ "<(PRODUCT_DIR)/<(module_name).node" ],
-          "destination": "<(module_path)"
-        }
-      ]
-    },
-    {
-      'target_name': 'node_printer',
+      "target_name": "node_printer",
       'sources': [
-        # is like "ls -1 src/*.cc", but gyp does not support direct patterns on
-        # sources
-        '<!@(["python", "tools/getSourceFiles.py", "src", "cc"])'
+        'src/node_printer.cc',
+        'src/node_printer_win.cc',
+        'src/node_printer_posix.cc'
       ],
       'include_dirs' : [
         "<!(node -e \"require('nan')\")"
