+++ examples/example_zebra_printer.js
@@ -1,4 +1,4 @@
-﻿var printer = require("../lib")
+﻿var printer = require("../printer")
 	, template = "N\nS4\nD15\nq400\nR\nB20,10,0,1,2,30,173,B,\"barcode\"\nP0\n";
 
 function printZebra(barcode_text, printer_name){
