+++ examples/getDefaultPrinterName.js
@@ -1,4 +1,4 @@
-var printer = require("../lib");
+var printer = require("../printer");
 
 console.log('default printer name: ' + (printer.getDefaultPrinterName() || 'is not defined on your computer'));
