+++ examples/getPrinters.js
@@ -1,3 +1,3 @@
-var printer = require("../lib"),
+var printer = require("../printer"),
     util = require('util');
 console.log("installed printers:\n"+util.inspect(printer.getPrinters(), {colors:true, depth:10}));
