+++ examples/getSupportedFormats.js
@@ -1,3 +1,3 @@
-var printer = require("../lib"),
+var printer = require("../printer"),
     util = require('util');
 console.log("supported formats are:\n"+util.inspect(printer.getSupportedPrintFormats(), {colors:true, depth:10}));
