+++ examples/getSupportedJobCommands.js
@@ -1,3 +1,3 @@
-var printer = require("../lib"), //=require('pritner')
+var printer = require("../printer"), //=require('pritner')
     util = require('util');
 console.log("supported job commands:\n"+util.inspect(printer.getSupportedJobCommands(), {colors:true, depth:10}));
