+++ examples/printPDFFileInBuffer.js
@@ -1,6 +1,6 @@
 // Windows does not support PDF formats, but you can use imagemagick-native to achieve conversion from PDF to EMF.
 
-var printer = require("../lib"),
+var printer = require("../printer"),
     fs = require('fs'),
     path = require('path'),
     filename = process.argv[2],
